testing complete for MPa and Kg to tensile ,compression and tear tests

+++ TYR_2.0_18.5/TY_03_REPORTS.py
@@ -2629,8 +2629,8 @@ class PlotCanvas(FigureCanvas):
                              ax.set_xlim(0,int(int(x[0])))
                              ax.set_ylim(0,int(x[1]))
                          else:
-                             ax.set_xlim(0,int(int(x[0])*0.1))
-                             ax.set_ylim(0,int(x[1]))
+                             ax.set_xlim(0,float(int(x[0])*0.1))
+                             ax.set_ylim(0,float(x[1]))
                 
                                           
                 connection.close
